+++ apps/www/pages/api/sponsors.ts
@@ -1,4 +1,4 @@
-import { NextApiRequest, NextApiResponse } from 'next'
+import type { NextApiRequest, NextApiResponse } from 'next'
 
 const AV_SIZE = 32
 const PADDING = 4