+++ packages/core/src/hooks/usePosition.ts
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
+import { autorun } from 'mobx'
 import * as React from 'react'
 import type { TLBounds } from '../types'
 
@@ -7,25 +8,30 @@ export function usePosition(bounds: TLBounds, rotation = 0) {
 
   // Update the transform
   React.useLayoutEffect(() => {
-    const elm = rBounds.current!
+    return autorun(() => {
+      const elm = rBounds.current!
 
-    const transform = `
+      const transform = `
     translate(
       calc(${bounds.minX}px - var(--tl-padding)),
       calc(${bounds.minY}px - var(--tl-padding))
     )
     rotate(${rotation + (bounds.rotation || 0)}rad)`
 
-    elm.style.setProperty('transform', transform)
+      elm.style.setProperty('transform', transform)
 
-    elm.style.setProperty('width', `calc(${Math.floor(bounds.width)}px + (var(--tl-padding) * 2))`)
+      elm.style.setProperty(
+        'width',
+        `calc(${Math.floor(bounds.width)}px + (var(--tl-padding) * 2))`
+      )
 
-    elm.style.setProperty(
-      'height',
-      `calc(${Math.floor(bounds.height)}px + (var(--tl-padding) * 2))`
-    )
+      elm.style.setProperty(
+        'height',
+        `calc(${Math.floor(bounds.height)}px + (var(--tl-padding) * 2))`
+      )
 
-    // elm.style.setProperty('z-index', zIndex + '')
+      // elm.style.setProperty('z-index', zIndex + '')
+    })
   }, [bounds, rotation])
 
   return rBounds