+++ packages/tldraw/src/state/TLDR.ts
@@ -15,6 +15,7 @@ import { Vec } from '@tldraw/vec'
 import type { TDShapeUtil } from './shapes/TDShapeUtil'
 import { getShapeUtil } from './shapes'
 
+const isDev = process.env.NODE_ENV === 'development'
 export class TLDR {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   static getShapeUtil<T extends TDShape>(type: T['type']): TDShapeUtil<T>
@@ -672,7 +673,7 @@ export class TLDR {
     const rotatedCenter = Vec.rotWith(center, origin, delta)
 
     // Get the top left point relative to the rotated center
-    const nextPoint = Vec.round(Vec.sub(rotatedCenter, relativeCenter))
+    const nextPoint = Vec.toFixed(Vec.sub(rotatedCenter, relativeCenter))
 
     // If the shape has handles, we need to rotate the handles instead
     // of rotating the shape. Shapes with handles should never be rotated,
@@ -685,7 +686,7 @@ export class TLDR {
           Object.entries(shape.handles).map(([handleId, handle]) => {
             // Rotate each handle's point around the shape's center
             // (in relative shape space, as the handle's point will be).
-            const point = Vec.round(Vec.rotWith(handle.point, relativeCenter, delta))
+            const point = Vec.toFixed(Vec.rotWith(handle.point, relativeCenter, delta))
             return [handleId, { ...handle, point }]
           })
         ) as T['handles'],
@@ -865,8 +866,8 @@ export class TLDR {
     return shapes.length === 0
       ? 1
       : shapes
-          .filter((shape) => shape.parentId === pageId)
-          .sort((a, b) => b.childIndex - a.childIndex)[0].childIndex + 1
+        .filter((shape) => shape.parentId === pageId)
+        .sort((a, b) => b.childIndex - a.childIndex)[0].childIndex + 1
   }
 
   /* -------------------------------------------------- */
@@ -891,4 +892,16 @@ export class TLDR {
       throw new Error()
     }
   }
+
+  static warn(e: any) {
+    if (isDev) {
+      console.warn(e);
+    }
+  }
+  static error(e: any) {
+    if (isDev) {
+      console.error(e)
+    }
+  }
+
 }