+++ packages/tldraw/src/state/shapes/TextUtil/TextUtil.tsx
@@ -11,6 +11,7 @@ import { Vec } from '@tldraw/vec'
 import { TLDR } from '~state/TLDR'
 import { getTextAlign } from '../shared/getTextAlign'
 import { getTextSvgElement } from '../shared/getTextSvgElement'
+import { stopPropagation } from '~components/stopPropagation'
 
 type T = TextShape
 type E = HTMLDivElement
@@ -23,6 +24,8 @@ export class TextUtil extends TDShapeUtil<T, E> {
   canEdit = true
 
   canBind = true
+
+  canClone = true
 
   getShape = (props: Partial<T>): T => {
     return Utils.deepMerge<T>(
@@ -91,7 +94,27 @@ export class TextUtil extends TDShapeUtil<T, E> {
 
       const handleKeyDown = React.useCallback(
         (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
-          e.stopPropagation()
+          // If this keydown was just the meta key or a shortcut
+          // that includes holding the meta key like (Command+V)
+          // then leave the event untouched. We also have to explicitly
+          // Implement undo/redo for some reason in order to get this working
+          // in the vscode extension. Without the below code the following doesn't work
+          //
+          // - You can't cut/copy/paste when when text-editing/focused
+          // - You can't undo/redo when when text-editing/focused
+          // - You can't use Command+A to select all the text, when when text-editing/focused
+          if (!(e.key === 'Meta' || e.metaKey)) {
+            e.stopPropagation()
+          } else if (e.key === 'z' && e.metaKey) {
+            if (e.shiftKey) {
+              document.execCommand('redo', false)
+            } else {
+              document.execCommand('undo', false)
+            }
+            e.stopPropagation()
+            e.preventDefault()
+            return
+          }
 
           if (e.key === 'Escape' || (e.key === 'Enter' && (e.ctrlKey || e.metaKey))) {
             e.currentTarget.blur()
@@ -199,9 +222,11 @@ export class TextUtil extends TDShapeUtil<T, E> {
                   onKeyDown={handleKeyDown}
                   onBlur={handleBlur}
                   onPointerDown={handlePointerDown}
+                  spellCheck="true"
                   wrap="off"
                   dir="auto"
                   datatype="wysiwyg"
+                  onContextMenu={stopPropagation}
                 />
               ) : (
                 text
@@ -285,7 +310,7 @@ export class TextUtil extends TDShapeUtil<T, E> {
     } = initialShape
 
     return {
-      point: Vec.round([bounds.minX, bounds.minY]),
+      point: Vec.toFixed([bounds.minX, bounds.minY]),
       style: {
         ...initialShape.style,
         scale: scale * Math.max(Math.abs(scaleY), Math.abs(scaleX)),
@@ -309,7 +334,7 @@ export class TextUtil extends TDShapeUtil<T, E> {
         ...shape.style,
         scale: 1,
       },
-      point: Vec.round(Vec.add(shape.point, Vec.sub(center, newCenter))),
+      point: Vec.toFixed(Vec.add(shape.point, Vec.sub(center, newCenter))),
     }
   }