+++ apps/vscode/extension/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.9
+
+- Updates READMEs.
+
+## 1.1.6
+
+- Fixes bugs in VS Code extension.
+
 ## 0.1.23
 
 - Fixing bugs related to saving files.